Completely rewrite everything
Replaces the old parser with the (heavily modified in this commit) AST
parser, adds a Compiler to turn the AST into opcodes, and updates the
matcher to use them. Also updates some of the public APIs.

There was no effort to break this up into small, atomic commits, as you
can probably tell. It's too early in the project to really care...